[#3] Comment out custom typeclasses
Problem: The custom typeclasses IsCommutativeRing and IsInteger are
likely inspired by the IsPeano and PeanoOrder typeclasses in
type-natural. However, the latter typeclasses are necessary to unify the
proofs for the built-in and library Nat types. There is no such
unification necessary for type-integers, so these typeclasses just add
unnecessary complexity at the moment.

Solution: Comment out the typeclasses, leaving open the possibility of
reintroducing them into the code if they become useful later.